refactor: move schema tests to colocated unit tests, remove evals dir
Schema validation tests belong as colocated unit tests in
karat/tui/label_app_test.py, not in tests/evals/. Deleted duplicates
that overlap with existing integration tests. Removed the custom
pytest eval runner -- agent evals should use skill-creator instead.
